requirements_test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements_test.in
#
airtable-python-wrapper==0.12.0  # via -r requirements.in
amqp==2.5.2               # via kombu
appdirs==1.4.4            # via black
atomicwrites==1.4.0       # via pytest
attrs==20.2.0             # via jsonschema, pytest
beautifulsoup4==4.9.3     # via directory-components
billiard==3.6.3.0         # via celery
black==20.8b1             # via -r requirements_test.in, blacken-docs
blacken-docs==1.6.0       # via -r requirements_test.in
boto3==1.10.23            # via -r requirements.in
botocore==1.13.50         # via boto3, s3transfer
celery[redis]==4.3.0      # via -r requirements.in
certifi==2020.6.20        # via elasticsearch, requests, sentry-sdk
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via black
codecov==2.1.7            # via -r requirements_test.in
coverage==5.3             # via codecov, pytest-cov
cryptography==3.2         # via -r requirements.in, pyopenssl, requests
dataclasses==0.8          # via black
defusedxml==0.6.0         # via odfpy
diff-match-patch==20200713  # via django-import-export
directory-client-core==6.3.0  # via directory-forms-api-client, directory-sso-api-client
directory-components==35.6.1  # via -r requirements.in
directory-constants==20.17.0  # via -r requirements.in, directory-components
directory-forms-api-client==5.1.0  # via -r requirements.in
directory-healthcheck==2.0.0  # via -r requirements.in
directory-sso-api-client==6.5.3  # via -r requirements.in
directory-validators==6.0.3  # via -r requirements.in
dj-database-url==0.5.0    # via -r requirements.in
django-admin-ip-restrictor==2.1.0  # via -r requirements.in
django-celery-beat==1.5.0  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-extensions==2.2.5  # via -r requirements.in
django-field-history==0.7.0  # via -r requirements.in
django-filter==2.2.0      # via -r requirements.in
django-health-check==3.8.0  # via directory-healthcheck
django-import-export==2.0.2  # via -r requirements.in
django-ipware==2.1.0      # via django-admin-ip-restrictor
django-redis==4.10.0      # via -r requirements.in
django-staff-sso-client==1.0.0  # via -r requirements.in
django-storages==1.8.0    # via -r requirements.in
django-timezone-field==4.0  # via django-celery-beat
django==2.2.13            # via -r requirements.in, directory-client-core, directory-components, directory-constants, directory-healthcheck, directory-validators, django-admin-ip-restrictor, django-filter, django-import-export, django-redis, django-staff-sso-client, django-storages, django-timezone-field, sigauth
django_json_widget==1.0.1  # via -r requirements.in
djangorestframework==3.10.3  # via -r requirements.in, sigauth
docutils==0.15.2          # via botocore
elasticsearch-dsl==7.3.0  # via -r requirements.in
elasticsearch==7.9.1      # via elasticsearch-dsl
et-xmlfile==1.0.1         # via openpyxl
factory-boy==2.12.0       # via -r requirements.in
faker==4.14.0             # via factory-boy
flake8==3.8.4             # via -r requirements_test.in
freezegun==0.3.12         # via -r requirements_test.in
future==0.18.2            # via django-json-widget
idna==2.8                 # via requests
importlib-metadata==2.0.0  # via flake8, jsonschema, kombu, pluggy, pytest
iso3166==1.0.1            # via -r requirements.in
isort==5.6.4              # via -r requirements_test.in
jdcal==1.4.1              # via openpyxl
jmespath==0.10.0          # via boto3, botocore
jsonschema==3.2.0         # via directory-components
kombu==4.6.6              # via -r requirements.in, celery
markdown2==2.3.9          # via -r requirements.in
markuppy==1.14            # via tablib
mccabe==0.6.1             # via flake8
mohawk==0.3.4             # via sigauth
monotonic==1.5            # via directory-client-core
more-itertools==8.5.0     # via pytest
mypy-extensions==0.4.3    # via black
numpy==1.19.2             # via pandas
oauthlib==3.1.0           # via requests-oauthlib
odfpy==1.4.1              # via tablib
olefile==0.46             # via directory-validators
openpyxl==3.0.5           # via tablib
packaging==20.4           # via pytest, pytest-sugar
pandas==1.0.1             # via -r requirements.in
pathspec==0.8.1           # via black
pillow==8.0.1             # via directory-validators
pluggy==0.13.1            # via pytest
pre-commit-hooks==3.3.0   # via -r requirements_test.in
psycopg2-binary==2.8.4    # via -r requirements.in
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.2.0           # via flake8
pyopenssl==19.1.0         # via requests
pyparsing==2.4.7          # via packaging
pyrsistent==0.17.3        # via jsonschema
pytest-cov==2.7.1         # via -r requirements_test.in
pytest-django==3.5.1      # via -r requirements_test.in
pytest-sugar==0.9.2       # via -r requirements_test.in
pytest==4.6.5             # via -r requirements_test.in, pytest-cov, pytest-django, pytest-sugar
python-crontab==2.5.1     # via django-celery-beat
python-dateutil==2.8.1    # via botocore, elasticsearch-dsl, faker, freezegun, pandas, python-crontab
pytz==2019.3              # via celery, directory-validators, django, django-timezone-field, pandas
pyyaml==5.3.1             # via tablib
raven==6.10.0             # via django-staff-sso-client
redis==3.5.3              # via celery, django-redis
regex==2020.11.13         # via black
requests-oauthlib==1.3.0  # via django-staff-sso-client
requests[security]==2.22.0  # via -r requirements.in, airtable-python-wrapper, codecov, directory-client-core, requests-mock, requests-oauthlib
requests_mock==1.6.0      # via -r requirements_test.in
ruamel.yaml.clib==0.2.2   # via ruamel.yaml
ruamel.yaml==0.16.12      # via pre-commit-hooks
s3transfer==0.2.1         # via boto3
sentry-sdk==0.13.4        # via -r requirements.in
sigauth==4.1.1            # via -r requirements.in, directory-client-core
six==1.15.0               # via airtable-python-wrapper, cryptography, django-extensions, elasticsearch-dsl, freezegun, jsonschema, mohawk, packaging, pyopenssl, pytest, python-dateutil, requests-mock, w3lib
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.4.1           # via django
tablib[html,ods,xls,xlsx,yaml]==2.0.0  # via django-import-export
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.3       # via faker
toml==0.10.2              # via black, pre-commit-hooks
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
urllib3==1.25.11          # via -r requirements.in, botocore, directory-validators, elasticsearch, requests, sentry-sdk
vine==1.3.0               # via amqp, celery
w3lib==1.22.0             # via directory-client-core
waitress==1.4.3           # via -r requirements.in
wcwidth==0.2.5            # via pytest
whitenoise==4.1.4         # via -r requirements.in
xlrd==1.2.0               # via tablib
xlwt==1.3.0               # via tablib
zipp==3.3.1               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
